RestDemo/requirements.txt
behave==1.2.6
behave-django==1.3.0
Django==2.2.10
djangorestframework==3.9.2
parse==1.14.0
parse-type==0.5.2
psycopg2-binary==2.7.7
Pygments==2.3.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
whitenoise==4.1.2
